refactor(log): extract `LevelFixedLogger` from `AdaptiveLogger`
